[CAS-335] Remove useless ChatCallImpl class
