+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api/ErrorCall.kt
@@ -1,10 +1,14 @@
 package io.getstream.chat.android.client.api
 
-import io.getstream.chat.android.client.call.ChatCallImpl
+import io.getstream.chat.android.client.call.Call
 import io.getstream.chat.android.client.errors.ChatError
 import io.getstream.chat.android.client.utils.Result
 
-internal class ErrorCall<T : Any>(private val e: ChatError) : ChatCallImpl<T>() {
+internal class ErrorCall<T : Any>(private val e: ChatError) : Call<T> {
+    override fun cancel() {
+        // Empty
+    }
+
     override fun execute(): Result<T> {
         return Result(null, e)
     }
